pyproject.toml
[project]
name = "curso-python-entornos-virtuales-pep8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.84.0",
    "groq>=1.0.0",
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.4",
]
